sql_optimizer_agent/requirements.txt
psycopg2-binary>=2.9.0
sqlparse>=0.4.0
openai>=1.0.0
rich>=13.0.0
